requirements.txt
fastapi
uvicorn
python-multipart
openpyxl
pywebview
pandas
requests
python-dotenv
